CLI: Move `plot` commands to separate top-level command
Move the `plot` commands, currently subcommands of the top-level launch
command, to a separate top-level command called `plot`. This is more
intuitive, since the `launch` commands (as the help describes) are for
"launching" the work chains.
